+++ src/main/java/combat/stats/Stats.java
@@ -4,23 +4,23 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Stats implements Cloneable{
-    private Map<Stat, Integer> stats;
+    private Map<Stat, Float> stats;
 
     public Stats(){
         stats = new HashMap<>();
         for(Stat stat : Stat.values())
-            stats.put(stat, 0);
+            stats.put(stat, 0f);
     }
 
-    public void put(Stat stat, Integer value){
+    public void put(Stat stat, Float value){
         stats.put(stat, value);
     }
 
-    public void addModifer(Stat stat, Integer value){
+    public void addModifer(Stat stat, Float value){
         stats.put(stat, stats.get(stat) + value);
     }
 
-    public Integer get(Stat stat){
+    public Float get(Stat stat){
         return stats.get(stat);
     }
